display, tap_tempo: carry two findings across the library extraction
Two fixes from the autonomous loop's standing PR (killroom-analog-multi-effect
#59) landed on files that have since moved into this library, so merging that
branch would have dropped them. Re-applied here, where the code now lives.

display::text_width accumulated the pen advance in a uint16_t and returned
uint8_t, and the fixed-font draw_text_right computed its pixel count the same
way, so a string past 255 px reported its width mod 256 -- a small number that
looks like it fits, which is worse than an obviously wrong one. No caller reaches
it on a 128 px panel; the type is what stops a longer string from quietly fooling
the right-align and font-fit callers. Both regression tests come across with it.

tap_tempo's header said MIDI clock takes priority over tap. It is the other way
round: a tap that measures a tempo clears the clock-sync flag, and the pedal
follows the clock again only once sync is re-armed by a preset load or the Sync
setting. The arbitration itself is tempo_controller's and was already correct.

+++ include/pedal_core/display.hpp
@@ -45,8 +45,10 @@ namespace display {
     void    draw_text(const Font& f, uint8_t x, uint8_t y, const char* str);
     // Like draw_text but cut off at x_max — including mid-glyph — to use every pixel.
     void    draw_text_clipped(const Font& f, uint8_t x, uint8_t y, const char* str, uint8_t x_max);
-    // Pixel width (sum of pen advances) the string would occupy.
-    uint8_t text_width(const Font& f, const char* str);
+    // Pixel width (sum of pen advances) the string would occupy. Returned as
+    // uint16_t so a string wider than 255 px reports its true width instead of
+    // wrapping mod 256, which would fool the right-align / font-fit callers.
+    uint16_t text_width(const Font& f, const char* str);
     // Draw a string ending its right edge at x_right (right-aligned).
     void    draw_text_right(const Font& f, uint8_t x_right, uint8_t y, const char* str);
